Made feature checks static
